// backend/src/types/ISeason.ts
export interface ISeason {
  currentSeasonNumber: number;
  league: string;
  total_goals_scored_home_last_season: number;
  total_goals_scored_away_last_season: number;
  last_season_home_mean_scored: number;
  last_season_away_mean_scored: number;
  last_season_home_mean_conceded: number;
  last_season_away_mean_conceded: number;
}